Add Package integration. Split program into different packages.
