+++ src/main/java/Duke.java
@@ -1,9 +1,27 @@
+import command.ByeCommand;
+import command.DeadlineCommand;
+import command.DeleteCommand;
+import command.EventCommand;
+import command.InvalidCommand;
+import command.ListCommand;
+import command.MarkCommand;
+import command.TodoCommand;
+import command.UnmarkCommand;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Scanner;
 import java.util.ArrayList;
+import objects.Deadline;
+import objects.DukeException;
+import objects.Event;
+import objects.Task;
+import objects.Todo;
+import parser.Parser;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
 
 /**
  * Duke is the main class for the application that runs according to the commands given to it by the user.
