+++ src/main/java/command/DeadlineCommand.java
@@ -1,8 +1,15 @@
+package command;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
+import objects.DukeException;
+import parser.Parser;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
 
 /**
  * The DeadlineCommand class handles the "deadline" command.
