+++ src/main/java/objects/DukeException.java
@@ -1,3 +1,8 @@
+package objects;
+
+/**
+ * The DukeException class represents exceptions specific to this application.
+ */
 public class DukeException extends Exception {
     private String exceptionMessage;
     public DukeException(String exceptionMessage) {
