+++ src/main/java/objects/Event.java
@@ -1,3 +1,5 @@
+package objects;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 public class Event extends Task{
