+++ src/main/java/objects/Task.java
@@ -1,3 +1,5 @@
+package objects;
+
 /**
  * The Task class represents the various different tasks in the task list of the application.
  */
